Libradtran: Allow ability to modify environment
If LibRadtran was compiled in a nonstandard way, you may need to modify
environment variables to make it run (e.g. load certain modules, or
modify `LD_LIBRARY_PATH`). This allows this to be done via the config
file, and basically just adds arbitrary code to the LUT bash script.